+++ internal/catalog/catalog_test.go
@@ -1,14 +1,19 @@
 package catalog_test
 
 import (
+	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"path/filepath"
 	"testing"
 
 	"github.com/arm/topo/internal/catalog"
+	"github.com/arm/topo/internal/deploy/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestListRepos(t *testing.T) {
+func TestParseTemplates(t *testing.T) {
 	t.Run("parses valid JSON successfully", func(t *testing.T) {
 		jsonData := []byte(`[
 			{
@@ -27,7 +32,7 @@ func TestListRepos(t *testing.T) {
 			}
 		]`)
 
-		templates, err := catalog.ParseRepos(jsonData)
+		templates, err := catalog.ParseTemplates(jsonData)
 
 		require.NoError(t, err)
 		assert.Len(t, templates, 2)
@@ -50,7 +55,7 @@ func TestListRepos(t *testing.T) {
 	t.Run("returns error for invalid JSON", func(t *testing.T) {
 		jsonData := []byte(`[{"id": "test", invalid}]`)
 
-		_, err := catalog.ParseRepos(jsonData)
+		_, err := catalog.ParseTemplates(jsonData)
 
 		require.Error(t, err)
 		assert.ErrorContains(t, err, "failed to unmarshal templates")
@@ -67,9 +72,52 @@ func TestListRepos(t *testing.T) {
 			}
 		]`)
 
-		_, err := catalog.ParseRepos(jsonData)
+		_, err := catalog.ParseTemplates(jsonData)
 
 		require.Error(t, err)
 		assert.ErrorContains(t, err, "failed to unmarshal templates")
+	})
+}
+
+func TestFetchTemplatesJSON(t *testing.T) {
+	t.Run("given a remote url, it fetches the catalog json", func(t *testing.T) {
+		catalogJSON := `[{"json": "for-real"}]`
+		server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			if r.URL.Path == "/give-json" {
+				w.Write([]byte(catalogJSON))
+			}
+		}))
+		defer server.Close()
+
+		url := fmt.Sprintf("%s/give-json", server.URL)
+		got, err := catalog.FetchTemplatesJSON(url)
+
+		require.NoError(t, err)
+		assert.Equal(t, catalogJSON, string(got))
+	})
+
+	t.Run("given a file:// url, it fetches the catalog json", func(t *testing.T) {
+		path := filepath.Join(t.TempDir(), "file.json")
+		catalogJSON := `[{"json": "actually-json"}]`
+		testutil.RequireWriteFile(t, path, catalogJSON)
+
+		url := fmt.Sprintf("file://%s", path)
+		got, err := catalog.FetchTemplatesJSON(url)
+
+		require.NoError(t, err)
+		assert.Equal(t, catalogJSON, string(got))
+	})
+
+	t.Run("returns error when request fails", func(t *testing.T) {
+		server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			http.NotFound(w, r)
+
+		}))
+		defer server.Close()
+
+		url := fmt.Sprintf("%s/give-json-pretty-please", server.URL)
+		_, err := catalog.FetchTemplatesJSON(url)
+
+		assert.Error(t, err)
 	})
 }
